Added Lightbox to Gallery
Added JS file and updated HTML and CSS to include lightbox... I cannot get it to work, Not sure why. But, I think part of the problem is that my gallery shows like it is inside of my video div. I've checked number of divs and it looks correct, but devtools show it as video on hover.